refactor: Rename app and script to jomp
Breaking: Rename modules and paths
