+++ jomp/ui.py
@@ -6,7 +6,8 @@ class AppUI(object):
         App.setObjectName("App")
         App.resize(940, 770)
         sizePolicy = QtWidgets.QSizePolicy(
-            QtWidgets.QSizePolicy.Policy.Expanding, QtWidgets.QSizePolicy.Policy.Expanding
+            QtWidgets.QSizePolicy.Policy.Expanding,
+            QtWidgets.QSizePolicy.Policy.Expanding,
         )
 
         sizePolicy.setHorizontalStretch(0)
@@ -18,9 +19,17 @@ class AppUI(object):
         self.canvas = QtWidgets.QGraphicsView(App)
         self.canvas.setGeometry(QtCore.QRect(10, 10, 611, 611))
         self.canvas.setMouseTracking(True)
-        self.canvas.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarPolicy.ScrollBarAlwaysOff)
-        self.canvas.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarPolicy.ScrollBarAlwaysOff)
-        self.canvas.setSizeAdjustPolicy(QtWidgets.QAbstractScrollArea.SizeAdjustPolicy.AdjustToContents)
+
+        self.canvas.setVerticalScrollBarPolicy(
+            QtCore.Qt.ScrollBarPolicy.ScrollBarAlwaysOff
+        )
+        self.canvas.setHorizontalScrollBarPolicy(
+            QtCore.Qt.ScrollBarPolicy.ScrollBarAlwaysOff
+        )
+        self.canvas.setSizeAdjustPolicy(
+            QtWidgets.QAbstractScrollArea.SizeAdjustPolicy.AdjustToContents
+        )
+
         self.canvas.setInteractive(True)
         self.canvas.setObjectName("canvas")
         self.verticalLayoutWidget = QtWidgets.QWidget(App)
@@ -67,24 +76,34 @@ class AppUI(object):
         font.setPointSize(10)
         self.info_text.setFont(font)
         self.info_text.setAlignment(
-            QtCore.Qt.AlignmentFlag.AlignLeading | QtCore.Qt.AlignmentFlag.AlignLeft | QtCore.Qt.AlignmentFlag.AlignTop
+            QtCore.Qt.AlignmentFlag.AlignLeading
+            | QtCore.Qt.AlignmentFlag.AlignLeft
+            | QtCore.Qt.AlignmentFlag.AlignTop
         )
 
         self.info_text.setWordWrap(True)
-        self.info_text.setTextInteractionFlags(QtCore.Qt.TextInteractionFlag.TextSelectableByMouse)
+        self.info_text.setTextInteractionFlags(
+            QtCore.Qt.TextInteractionFlag.TextSelectableByMouse
+        )
+
         self.info_text.setObjectName("info_text")
         self.path_text = QtWidgets.QLabel(App)
         self.path_text.setGeometry(QtCore.QRect(630, 270, 301, 201))
+
         font = QtGui.QFont()
         font.setPointSize(10)
         self.path_text.setFont(font)
         self.path_text.setAlignment(
-            QtCore.Qt.AlignmentFlag.AlignBottom | QtCore.Qt.AlignmentFlag.AlignLeading | QtCore.Qt.AlignmentFlag.AlignLeft
+            QtCore.Qt.AlignmentFlag.AlignBottom
+            | QtCore.Qt.AlignmentFlag.AlignLeading
+            | QtCore.Qt.AlignmentFlag.AlignLeft
         )
 
         self.path_text.setWordWrap(True)
-        self.path_text.setTextInteractionFlags(QtCore.Qt.TextInteractionFlag.TextSelectableByMouse)
         self.path_text.setObjectName("path_text")
+        self.path_text.setTextInteractionFlags(
+            QtCore.Qt.TextInteractionFlag.TextSelectableByMouse
+        )
 
         self.retranslateUi(App)
         QtCore.QMetaObject.connectSlotsByName(App)
